Made some progress.
Clicking on parts of file triggers event.
Wrote a script that can be attached to any game object, which when
clicked (can be changed to when dragged and dropped) effects the stats
of the suspect.
Wrote a script that highlights whatever gameobject it's attached to when
moused over.
There's a tempFileOnDesk, when clicked it opens the current suspects
file.  If already open it will close it.  The file can also be closed by
clicking the red X in the upper left corner.