Changed a lot of stuff.
Got the dialog to properly display and disappear.
Now we can have conversations by populating a dictionary and calling the
Dialog method inside the TypeWriter class.  Refer to Suspect1Events for
an example.
The drag and drop needs some tweaking.  I'm not sure how to catch when
the drop occurs on the drawer.  All that's happening as of now is a
constant collision when the tool is over the drawer.  The stats are
being modified (more than once, will be fixed when drag and drop is
fixed) correctly.
May have missed some minor changes.