Read the description
Removed extra File # 1 prefab and random FileOpen prefab.

Modified LoadScene to allow execution from Button.

Changed ModifyStat to look better from the Scene Builder.

Added a few lines of code to DialogQueuer.  Can change the
lettersPerSecond from scene builder on this script.

If you want dialog to display when you drop a tool in the box, just add
DialogQueuer to the tool, but don't add Button.  Same goes for
ModifyStat.

Tools can only be taken off the toolRack when it is pulled out.

Modified drag to contain the original location of the tool, when the
toolRack is pulled out.

Can drag a tool into the drawer.  This will modify the stats of the
suspect, and type any dialog attached to the tool.  Can be modified to
do other things when needed.

Tools, if not dropped in the drawer, snap back to their original
location in the tool rack.