Tools Highlight, Label Creation
Created a system to highlight the tools when hovered over and display a
label. The Label should vanish over time if not picked up and the mouse
is no longer hovering over it and appear over time if the mouse is over
it. The label completely vanishes instantly if picked up.