Created Timer That Properly Displays A Countdown
Timer now contains a script called Time Manager.  This allows
specification of a start time.
There's also a variable for interval.  This controls how many real
seconds are treated as a single second in the game.
An interval of 5 means 5 real seconds, are treated as a single second in
game.
An interval of .1 means 1 real second is treated as 10 seconds in game.
Default is at .5 right now.

Also created a SubtractTime script that implements ButtonAction.  I
currently have one setup on the far left Number as an example.  It
subtracts 25 minutes currently.