Read My Description
I modified Drag and MoveToTools to disable to user from dragging a tool
when the toolRack is still moving.  Also made it so the user cannot drag
or open the toolRack when the file is open.

Added FadeInFadeOut script.  Properly fades labels in and out when
moused over, dragged, etc.