DialogForks is working now.  Read my description.
DialogForks is working now.  Each time an object is clicked it queues a
random DialogQueuer which then can't be queued anymore.  The
DialogQueuer can be marked as containing a key.
DialogForks can be used in place of DialogQueuer, just have the array
contain one element.

Modified TextScroller and DialogQueuer to set components as disabled and
not the entire gameObject.  Was having issues with it not existing when
disabled.