Read Notes
All the UI is currently working with the exception of a few things that
still need to be fixed.
..The MoveToTools Script is now broken. The UI has a seperate coordinate
system and the script will need to be changed.
..The Fading tools no longer work. I believe it's because the labels are
now sprites and no longer 3d objects with materials. Take a look

STORY IS COMING. I'm sorry, haven't had much time over the past few
days. Will be devoting the better part of tomorrow working on it.
Therefore, when working on stuff copy the interrogation scene and work
on that if you'd like to test things out there.