Read Description
Adapted MoveToTools for UI, kind of.  The Canvas makes things act oddly.
For placeToMove, a 0 moves the toolBox to the center of the screen.  35
is the default right now and is approximately what it was before.  I
have location destination and placeToMove as public but only the
placeToMove should be set by the user in the scene builder.

Fixed bug where the first person to talk was Harvey on occasion.

Should now be using only DialogForks (even if it's just size 1) instead
of DialogQueuer

Reordered hierarchy to make File appear before ToolRack.