Read Description
Dialog Plays on Start,  can be disabled on TheGameManager

Tools are inactive until all keys have been discovered

An event plays when all keys are discovered

The map is blank until a location has been discovered.

When a location is discovered the map is automatically pulled out, the
location fades in on the map, then the map is automatically pulled back
in.

Can click the map to pull out, just like the toolRack