Read Description
Keys, Locations, and Hints are now triggered on the line of Dialog and
not on the Sequence of dialog.

When keys or hints are found the New Hint Found pops up and disappears,
same with location but New Location Discovered pops up.

When a location is discovered it's unlocked and clickable.

Support was added that fixes most of the bugs introduced by the restart
button.  It's there if we want it later.

Started implementing the notepad and its pages.  Done for today though,
will continue tomorrow.